Alerta de confirmación para eliminar reseña y rol en perfil funcional

+++ src/app/components/pages/movie/movie.component.ts
@@ -205,26 +205,39 @@ export class MovieComponent {
   }
 
   deleteReview(reviewId: string) {
-    this.reviewsService.deleteReview(reviewId, this.movie.id).subscribe({
-      next: (_response) => {
-        Swal.fire({
-          title: 'Reseña eliminada',
-          text: 'Tu reseña ha sido eliminada con éxito.',
-          icon: 'success',
-          showCloseButton: true
-        }).then(() => {
-          delete this.reviews[reviewId];
-        });
-      }, 
-      error: (error) => {
-        console.log(error)
-        Swal.fire({
-          title: 'Error',
-          text: 'Hubo un problema al eliminar tu reseña. Por favor, intenta de nuevo.',
-          icon: 'error',
-          showCloseButton: true
-        });
-      }
-    })
+    Swal.fire({
+        title: '¿Estás seguro?',
+        text: 'No podrás revertir esta acción.',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Sí, eliminar',
+        cancelButtonText: 'Cancelar'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            this.reviewsService.deleteReview(reviewId, this.movie.id).subscribe({
+                next: (_response) => {
+                    Swal.fire({
+                        title: 'Reseña eliminada',
+                        text: 'Tu reseña ha sido eliminada con éxito.',
+                        icon: 'success',
+                        showCloseButton: true
+                    }).then(() => {
+                        delete this.reviews[reviewId];
+                    });
+                },
+                error: (error) => {
+                    console.log(error);
+                    Swal.fire({
+                        title: 'Error',
+                        text: 'Hubo un problema al eliminar tu reseña. Por favor, intenta de nuevo.',
+                        icon: 'error',
+                        showCloseButton: true
+                    });
+                }
+            });
+        }
+    });
   }
 }
